requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.2
python-docx==1.2.0
requests==2.33.1
streamlit==1.56.0